moves template pdf functions to new file
